addressed feedback from pull request. moved course photo urls to constants
