+++ utils/constants/index.js
@@ -1,5 +1,14 @@
 // export all constants from this file
 
-const defaultProfilePic = "https://www.vhv.rs/dpng/d/426-4264903_user-avatar-png-picture-avatar-profile-dummy-transparent.png"
+const defaultProfilePic = "https://www.vhv.rs/dpng/d/426-4264903_user-avatar-png-picture-avatar-profile-dummy-transparent.png";
 
-export default defaultProfilePic;
+const coursePhotos = {
+  'Science': 'https://epi-rsc.rsc-cdn.org/globalassets/05-journals-books-databases/our-journals/00-journal-pages-heros/Chemical-Science-HERO.jpg?version=9e72b3c3',
+  'History': 'https://alameda.edu/wp-content/uploads/2021/07/History.png',
+  'Literature': 'https://media.istockphoto.com/photos/book-and-glowing-letters-picture-id522513933?k=20&m=522513933&s=612x612&w=0&h=NjqmZlnwoTM7wgLI_rx1hiGdLb2xwT9UixPYR2d1ciI=',
+  'Language': 'https://i0.wp.com/www.michigandaily.com/wp-content/uploads/2021/03/0-1.jpg?fit=1200%2C960&ssl=1',
+  'Math': 'https://www.actuaries.org.uk/sites/default/files/curriculum/Actuarial%20Mathematics%20%28CM1%29%20Banner.jpg'
+}
+
+// export default defaultProfilePic;
+module.exports = { defaultProfilePic, coursePhotos }
